+++ components/admin/TextEditor/config.js
@@ -41,11 +41,6 @@ export const config = {
     ],
   },
   language: 'en',
-  mediaEmbed: {
-    removeProviders: [
-      'dailymotion', 'spotify', 'youtube', 'vimeo',
-    ],
-  },
   toolbar: {
     items: [
       'heading',